FlexAPI Initial Commit
FlexAPI is a new API for the Flexlay editor, to form Flexlay 1.0.0 in the future.
This is needed to ensure that everything is documented, and code
which uses the api is easier to understand. I also hope to provide more
functionality than Flexlay alone currently provides, in the long term.
This commit adds the very simplest of features to the new api,
including the menu bar and a text editor.

Multiple commits condensed due to human error using git.